==> 0_data_cleaning.ipynb <==
import pandas as pd
import re
from datetime import datetime

events = pd.read_csv("data/Dataset1.csv")
accidents = pd.read_csv("data/Dataset2.csv")

### Fixing Columns

events.rename(str.lower, axis='columns', inplace=True)
accidents.rename(str.lower, axis='columns', inplace=True)
events.columns = events.columns.str.replace("\s", '_', regex=True).str.replace("\W", '', regex=True)
events.rename(columns=lambda x: x[re.search("[a-zA-Z]", x).start():], inplace=True)
accidents.columns = accidents.columns.str.replace("\s", '_', regex=True).str.replace("\W", '', regex=True)
accidents.rename(columns=lambda x: x[re.search("[a-zA-Z]", x).start():], inplace=True)

# primary occupation code = North American Industry Classification System codes (NAICS)
# National Occupational Classification (NOC)
events = events.rename(columns={'event_date':'date', 'municipality':'city', 'primary_occupation_code':'NAICS'})
accidents = accidents.rename(columns={'accident_date':'date', 'organization_city':'city', 'occupation_code':'NOC'})

### Date Field

events['date'] = pd.to_datetime(events['date'], format="mixed")
events = events[events['date'] < datetime.now()]

accidents['date'] = pd.to_datetime(accidents['date'], errors='coerce')
accidents = accidents.dropna(subset=['date']).reset_index(drop=True)
accidents = accidents[accidents['date'] < "2024-12-31"]

### Cleaning Text Fields

def fix_text(s:pd.Series)->pd.Series:
    s = s.str.replace("\s", '_', regex=True)
    s = s.str.replace("\W", '', regex=True)
    s.fillna('', inplace=True)
    s = s.astype(str)
    s = s.apply(str.lower)
    return s

events['city'] = fix_text(events['city'])

print(list(events['city'].unique()))

accidents['city'] = fix_text(accidents['city'])
accidents['industry_sector_description'] = fix_text(accidents['industry_sector_description'])
print(list(accidents['industry_sector_description'].unique()))

#### Occupation Description

events['occupation_description'] = events['occupation_description'].str.lower()
accidents['occupation_description'] = accidents['occupation_description'].str.lower()

#### NOC

accidents['NOC'] = accidents['NOC'].apply(lambda x: pd.NA if re.findall("\D", x) else x)
# accidents = accidents.dropna(subset=['NOC'])
accidents

### Age

accidents = accidents[accidents['worker_age'] <= 100]
accidents['worker_experience_in_years'] = accidents['worker_experience_in_years'].fillna(0)
accidents = accidents[accidents['worker_age'] >= accidents['worker_experience_in_years']]
accidents

### Saving

events.to_csv('cleaned_data/events.csv', index=0)
accidents.to_csv('cleaned_data/accidents.csv', index=0)
