==> 0_merging.ipynb <==
import pandas as pd
from thefuzz import process
# from fuzzywuzzy import process

events = pd.read_csv('cleaned_data/events.csv')
accidents = pd.read_csv('cleaned_data/accidents.csv')

events['sector'] = events['occupation_description'] + " " + events['sector_description']

sector_description_1 = events['sector'].dropna().unique()
sector_description_2 = (accidents['occupation_description'] + " " + accidents['industry_sector_description'].str.replace('_', ' ')).dropna().unique()

# fuzzy matching
fuzzy_matches = []
for desc2 in sector_description_2:
    match, score = process.extractOne(desc2, sector_description_1)
    fuzzy_matches.append((desc2, match, score))

fuzzy_match_df = pd.DataFrame(fuzzy_matches, columns=['industry_sector_description', 'matched_sector_description', 'match_score'])

fuzzy_match_df

#### Dates

events['date'] = pd.to_datetime(events['date'], format="%Y-%m-%d")
accidents['date'] = pd.to_datetime(accidents['date'], format="%Y-%m-%d")

### Replacing industry_sector_description with matched_sector_description in accidents dataset

accidents['industry_sector_description'] = accidents['occupation_description'] + " " + accidents['industry_sector_description'].str.replace('_', ' ')
accidents = pd.merge(accidents, fuzzy_match_df[['industry_sector_description', 'matched_sector_description']], how='left', on='industry_sector_description')
accidents = accidents.drop(columns=['occupation_description', 'industry_sector_description']).rename(columns={'matched_sector_description':'sector'})
accidents

### Merging with events dataset

# events['sector'] = events['occupation_description'] + " " + events['sector_description']
# events = events.drop(columns=['occupation_description'])

temp = events.copy()
# temp['date'] = temp['date'].dt.strftime("%Y-%m")
temp['year'] = temp['date'].dt.year
temp['month'] = temp['date'].dt.month
accidents['year'] = accidents['date'].dt.year
accidents['month'] = accidents['date'].dt.month
temp = temp.drop(columns=['date'])
temp = temp.drop_duplicates()
merged = pd.merge(accidents, temp, on=['year', 'month', 'city', 'sector'], how='inner')
merged

#### Worker Count

import re
def text2num(item:str):
    res = re.findall("\d+", item)
    if len(res) == 1:
        return int(res[0])
    else:
        return (int(res[0]) + int(res[1]))//2

merged = merged.dropna(subset='worker_count')
merged['avg_worker_count'] = merged['worker_count'].apply(text2num)
merged

#### Saving

merged.to_csv("cleaned_data/merged.csv", index=0)

temp = merged.copy()
temp['date'] = temp['date'].dt.year
temp[['date', 'sector', 'worker_count', 'avg_worker_count']].drop_duplicates().head(20)
